+++ scripts/installers/mac_install.sh
@@ -18,19 +18,18 @@ then
 fi
 
 echo "Setting up conda environment…"
-conda create -n $CONDA_ENV_NAME -c conda-forge compas -y &> /dev/null
+conda create -n $CONDA_ENV_NAME python=3.10 -y &> /dev/null
 echo "Setting up conda environment…Done"
 
 echo "Activating environment…"
 conda activate $CONDA_ENV_NAME
 echo "Activating environment…Done!"
 
-echo "Installing compas_future…"
-python -m pip install --force-reinstall --no-input --quiet ./compas_future-main.zip
-python -m compas_rhino.install
-echo "Installing compas_future…Done!"
+echo "Installing compas from main..."
+python -m pip install --no-input --quiet compas@git+https://github.com/compas-dev/compas@main
+echo "Installing compas from main...Done!"
 
 echo "Installing compas_timber…"
-python -m pip install --force-reinstall --no-input --quiet ./compas_timber-dev.zip
+python -m pip install --force-reinstall --no-input --quiet compas_timber
 python -m compas_rhino.install
 echo "Installing compas_timber…Done!"