+++ src/compas_timber/connections/l_butt.py
@@ -9,15 +9,44 @@ from .solver import JointTopology
 
 
 class LButtJoint(Joint):
+    """Represents an L-Butt type joint which joins two beam in their ends, trimming the main beam.
+
+    This joint type is compatible with beams in L topology.
+
+    Please use `LButtJoint.create()` to properly create an instance of this class and associate it with an assembly.
+
+    Parameters
+    ----------
+    assembly : :class:`~compas_timber.assembly.TimberAssembly`
+        The assembly associated with the beams to be joined.
+    main_beam : :class:`~compas_timber.parts.Beam`
+        The main beam to be joined.
+    cross_beam : :class:`~compas_timber.parts.Beam`
+        The cross beam to be joined.
+
+    Attributes
+    ----------
+    beams : list(:class:`~compas_timber.parts.Beam`)
+        The beams joined by this joint.
+    cutting_plane_main : :class:`~compas.geometry.Frame`
+        The frame by which the main beam is trimmed.
+    cutting_plane_cross : :class:`~compas.geometry.Frame`
+        The frame by which the cross beam is trimmed.
+    joint_type : str
+        A string representation of this joint's type.
+
+
+    """
+
     SUPPORTED_TOPOLOGY = JointTopology.TOPO_L
 
-    def __init__(self, assembly=None, main_beam=None, cross_beam=None):
-        super(LButtJoint, self).__init__(assembly, [main_beam, cross_beam])
-        self.main_beam_key = main_beam.key
-        self.cross_beam_key = cross_beam.key
+    def __init__(self, main_beam=None, cross_beam=None, gap=0.0, frame=None, key=None):
+        super(LButtJoint, self).__init__(frame=frame, key=key)
         self.main_beam = main_beam
         self.cross_beam = cross_beam
-        self.gap = 0.0  # float, additional gap, e.g. for glue
+        self.main_beam_key = main_beam.key if main_beam else None
+        self.cross_beam_key = cross_beam.key if cross_beam else None
+        self.gap = gap  # float, additional gap, e.g. for glue
         self.features = []
 
     @property
@@ -30,20 +59,16 @@ class LButtJoint(Joint):
         data_dict.update(super(LButtJoint, self).data)
         return data_dict
 
-    @data.setter
-    def data(self, value):
-        Joint.data.fset(self, value)
-        self.main_beam_key = value["main_beam_key"]
-        self.cross_beam_key = value["cross_beam_key"]
-        self.gap = value["gap"]
+    @classmethod
+    def from_data(cls, value):
+        instance = cls(frame=Frame.from_data(value["frame"]), key=value["key"], gap=value["gap"])
+        instance.main_beam_key = value["main_beam_key"]
+        instance.cross_beam_key = value["cross_beam_key"]
+        return instance
 
     @property
     def beams(self):
         return [self.main_beam, self.cross_beam]
-
-    def restore_beams_from_keys(self, assemly):
-        self.main_beam = assemly.find_by_key(self.main_beam_key)
-        self.cross_beam = assemly.find_by_key(self.cross_beam_key)
 
     @property
     def joint_type(self):
@@ -62,10 +87,15 @@ class LButtJoint(Joint):
         cfr = max(angles_faces, key=lambda x: x[0])[1]
         return cfr
 
+    def restore_beams_from_keys(self, assemly):
+        """After de-serialization, resotres references to the main and cross beams saved in the assembly."""
+        self.main_beam = assemly.find_by_key(self.main_beam_key)
+        self.cross_beam = assemly.find_by_key(self.cross_beam_key)
+
     def add_features(self):
-        """
-        Adds the feature definitions (geometry, operation) to the involved beams.
-        In a T-Butt joint, adds the trimming plane to the main beam (no features for the cross beam).
+        """Adds the required extension and trimming features to both beams.
+
+        This method is automatically called when joint is created by the call to `Joint.create()`.
 
         """
         if self.features: