+++ src/compas_timber/ghpython/components/CT_Attributes_Get_Custom/code.py
@@ -6,14 +6,14 @@ from compas_timber.ghpython.rhino_object_name_attributes import get_obj_attribut
 
 
 class Attributes_Get_Custom(component):
-    def RunScript(self, refCrv):
-        if not refCrv:
-            self.AddRuntimeMessage(Warning, "Input parameter refCrv failed to collect data")
+    def RunScript(self, RefCrv):
+        if not RefCrv:
+            self.AddRuntimeMessage(Warning, "Input parameter RefCrv failed to collect data")
 
         AttributeName = []
         AttributeValue = []
 
-        guid = refCrv
+        guid = RefCrv
         if guid:
             attrdict = get_obj_attributes(guid)
             if attrdict: