+++ src/compas_timber/parts/features.py
@@ -1,8 +1,8 @@
 from compas.geometry import Brep
 from compas.geometry import BrepTrimmingError
 from compas.geometry import Frame
-from compas_future.datastructures import GeometricFeature
-from compas_future.datastructures import ParametricFeature
+from compas.datastructures import GeometricFeature
+from compas.datastructures import ParametricFeature
 
 
 def _trim_brep_with_frame(brep, frame):
@@ -40,17 +40,17 @@ class BeamTrimmingFeature(GeometricFeature):
         self._geometry = Frame.from_data(value["trimming_frame"])
 
     def apply(self, part):
-        part_geometry = part.get_geometry(with_features=True)
+        part_geometry = part.geometry
         if not isinstance(part_geometry, Brep):
-            raise ValueError("Brep feature {} cannot be applied to part with non Brep geometry {}".format(self, part))
-        g_copy = part_geometry.copy()
+            raise ValueError(
+                "Brep feature {} cannot be applied to part with geometry of type:{}".format(self, type(part_geometry))
+            )
         operation = self.OPERATIONS[Brep]
-
         try:
-            operation(g_copy, self._geometry)
+            operation(part_geometry, self._geometry)
+            return True, part_geometry
         except BrepTrimmingError:
-            False, part_geometry
-        return True, g_copy
+            return False, part.geometry  # copy again since operation is in-place and brep might be corrupted
 
     def __repr__(self):
         return "{}({})".format(self.__class__.__name__, repr(self._geometry))
@@ -79,9 +79,11 @@ class BeamBooleanSubtraction(GeometricFeature):
         return BeamBooleanSubtraction(self._geometry.copy())
 
     def apply(self, part):
-        part_geometry = part.get_geometry(with_features=True)
+        part_geometry = part.geometry
         if not isinstance(part_geometry, Brep):
-            raise ValueError("Brep feature {} cannot be applied to part with non Brep geometry {}".format(self, part))
+            raise ValueError(
+                "Brep feature {} cannot be applied to part with non Brep geometry {}".format(self, part_geometry)
+            )
         operation = self.OPERATIONS[Brep]
         try:
             return True, operation(part_geometry, self._geometry)
@@ -115,11 +117,28 @@ class BeamExtensionFeature(ParametricFeature):
         part.extend_ends(-self._extend_start, -self._extend_end)
 
     def accumulate(self, feature):
+        """Returns a new BeamExtensionFeature which the accumulative effect of this and `feature`.
+
+        Parameters
+        ----------
+        feature: :class:`compas_timber.features.BeamExtensionFeature`
+            The feature to accumulate with.
+
+        Returns
+        -------
+        :class:`~compas_timber.features.BeamExtensionFeature`
+            A new instance of BeamExtensionFeature.
+
+        """
         if not isinstance(feature, self.__class__):
-            return False
-        self._extend_start = max(feature._extend_start)
-        self._extend_end = max(feature._extend_end)
-        return True
+            raise TypeError(
+                "This feature {} cannot be accumulated with feature of type: {}".format(
+                    self.__class__.__name__, feature.__class__.__name__
+                )
+            )
+        return BeamExtensionFeature(
+            max(self._extend_start, feature._extend_start), max(self._extend_end, feature._extend_end)
+        )
 
     def __repr__(self):
         return "{}({}, {})".format(self.__class__.__name__, self._extend_start, self._extend_end)