+++ tests/compas_timber/test_joint.py
@@ -9,7 +9,7 @@ from compas.geometry import Point
 from compas.geometry import Vector
 
 from compas_timber.assembly import TimberAssembly
-from compas_timber.connections import Joint
+from compas_timber.connections import TButtJoint
 from compas_timber.connections import find_neighboring_beams
 from compas_timber.parts import Beam
 
@@ -38,7 +38,7 @@ def test_create(mocker):
     B2 = Beam(Frame.worldYZ(), length=1.0, width=0.1, height=0.1, geometry_type="mesh")
     A.add_beam(B1)
     A.add_beam(B2)
-    J = Joint.create(A, B1, B2)
+    J = TButtJoint.create(A, B1, B2)
 
     assert len(list(A.graph.nodes())) == 3
     assert len(list(A.graph.edges())) == 2
@@ -54,7 +54,7 @@ def test_joint_override_protection(mocker):
     A.add_beam(B1)
     A.add_beam(B2)
     A.add_beam(B3)
-    J = Joint.create(A, B1, B2)
+    J = TButtJoint.create(A, B1, B2)
 
     assert A.are_parts_joined([B1, B2])
     assert A.are_parts_joined([B1, B3]) is False
@@ -74,7 +74,7 @@ def test_deepcopy(mocker):
     B2 = Beam.from_endpoints(Point(1, 0, 0), Point(1, 1, 0), 0.1, 0.2, z_vector=Vector(0, 0, 1), geometry_type="mesh")
     A.add_beam(B1)
     A.add_beam(B2)
-    J = Joint(A, B1, B2)
+    J = TButtJoint.create(A, B1, B2)
     J_copy = deepcopy(J)
 
     assert J_copy is not J
@@ -84,16 +84,16 @@ if not compas.IPY:
 
     def test_find_neighbors(example_beams):
         expected_result = [
-            set([0, 3]),
-            set([0, 6]),
             set([0, 1]),
-            set([0, 5]),
+            set([0, 3]),
             set([1, 2]),
             set([1, 4]),
+            set([1, 5]),
             set([2, 3]),
             set([2, 6]),
-            set([2, 5]),
-            set([4, 5]),
+            set([3, 4]),
+            set([3, 5]),
+            set([5, 6]),
         ]
         result = find_neighboring_beams(example_beams)
         # beam objects => sets of keys for easy comparison